+++ src/presentation/controllers/signup.controller.ts
@@ -1,5 +1,4 @@
-import { MissingParamError } from '../errors/missing-param-error'
-import { InvalidParamError } from '../errors/invalid-param-error'
+import { MissingParamError, InvalidParamError } from '../errors'
 import { badRequest, serverError } from '../helpers/http-helper'
 import { type Controller } from '../protocols/controller'
 import { type IHttpRequest, type IHttpResponse } from '../protocols/http'
